// services/personalize-service/src/types.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface Payload {
  event_type: string
  event_timestamp: number
  arrival_timestamp: number
  event_version: string
  application: Application
  client: Client
  device: Device
  session: Session
  attributes: Attributes
  awsAccountId: string
}

export interface Application {
  app_id: string
  cognito_identity_pool_id: string
  package_name: string
  sdk: SDK
  title: string
  version_code: string
}

export interface SDK {
  name: string
  version: string
}

export interface Attributes {
  news_id: string
  user_id: string
}

export interface Client {
  client_id: string
  cognito_id: string
}

export interface Device {
  platform: Platform
}

export interface Platform {}

export interface Session {
  session_id: string
  start_timestamp: number
}